# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Priva Job Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user's profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Create or update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/{userId}:
    get:
      operationId: getProfileByUserId
      tags: [profile]
      summary: Get profile by user ID (recruiter/admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found

  # Companies
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List/search jobs
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: workMode
          in: query
          schema:
            type: string
            enum: [remote, hybrid, onsite]
        - name: employmentType
          in: query
          schema:
            type: string
            enum: [full-time, part-time]
        - name: experienceLevel
          in: query
          schema:
            type: string
            enum: [fresher, "1-3", "3-5", "5+"]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, pending]
        - name: companyId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new job (recruiter)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jobs/{id}/status:
    patch:
      operationId: updateJobStatus
      tags: [jobs]
      summary: Update job status (admin approve/reject)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobStatusChange"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

  # Applications
...
  /saved-jobs:
    get:
      operationId: listSavedJobs
      tags: [jobs]
      summary: List saved jobs for the current user
      responses:
        "200":
          description: List of saved jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedJob"
        "401":
          description: Not authenticated
    post:
      operationId: saveJob
      tags: [jobs]
      summary: Save a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveJobInput"
      responses:
        "201":
          description: Job saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedJob"
        "400":
          description: Already saved or invalid input
        "401":
          description: Not authenticated
  /saved-jobs/{jobId}:
    delete:
      operationId: unsaveJob
      tags: [jobs]
      summary: Remove a saved job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Unsaved
        "401":
          description: Not authenticated
        "404":
          description: Not found

  # Admin
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [job_seeker, recruiter, admin]
        createdAt:
          type: string
          format: date-time

    # Profile schemas
    ProfileInput:
      type: object
      properties:
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        education:
          type: ["string", "null"]
        experienceYears:
          type: ["integer", "null"]
        expectedSalary:
          type: ["integer", "null"]
        resumeUrl:
          type: ["string", "null"]
    Profile:
      type: object
      required: [id, userId]
      properties:
        id:
          type: integer
        userId:
          type: integer
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        education:
          type: ["string", "null"]
        experienceYears:
          type: ["integer", "null"]
        expectedSalary:
          type: ["integer", "null"]
        resumeUrl:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/User"

    # Company schemas
...
    Company:
      type: object
      required: [id, recruiterId, name, createdAt]
      properties:
        id:
          type: integer
        recruiterId:
          type: integer
        name:
          type: string
        website:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logo:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # Job schemas
    JobInput:
      type: object
      required: [title, description, location, employmentType, workMode]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        experienceMin:
          type: ["integer", "null"]
        experienceMax:
          type: ["integer", "null"]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        workMode:
          type: string
          enum: [remote, hybrid, onsite]
        employmentType:
          type: string
          enum: [full-time, part-time]
        companyId:
          type: ["integer", "null"]
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        experienceMin:
          type: ["integer", "null"]
        experienceMax:
          type: ["integer", "null"]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        workMode:
          type: string
          enum: [remote, hybrid, onsite]
        employmentType:
          type: string
          enum: [full-time, part-time]
    JobStatusChange:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, inactive, pending]
    Job:
      type: object
      required: [id, title, description, location, employmentType, workMode, status, postedAt]
      properties:
        id:
          type: integer
        companyId:
          type: ["integer", "null"]
        recruiterId:
          type: integer
        title:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        experienceMin:
          type: ["integer", "null"]
        experienceMax:
          type: ["integer", "null"]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        workMode:
          type: string
          enum: [remote, hybrid, onsite]
        employmentType:
          type: string
          enum: [full-time, part-time]
        status:
          type: string
          enum: [active, inactive, pending]
        postedAt:
          type: string
          format: date-time
        company:
          $ref: "#/components/schemas/Company"
    JobListResponse:
      type: object
      required: [jobs, total]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
...
    SaveJobInput:
      type: object
      required: [jobId]
      properties:
        jobId:
          type: integer
    SavedJob:
      type: object
      required: [id, userId, jobId, savedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        jobId:
          type: integer
        savedAt:
          type: string
          format: date-time
        job:
          $ref: "#/components/schemas/Job"
...