# lib/api-zod/src/generated/types/jobUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Priva Job Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { JobUpdateEmploymentType } from './jobUpdateEmploymentType';
import type { JobUpdateWorkMode } from './jobUpdateWorkMode';

export interface JobUpdate {
  title?: string;
  description?: string;
  skills?: string[];
  location?: string;
  /** @nullable */
  experienceMin?: number | null;
  /** @nullable */
  experienceMax?: number | null;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  workMode?: JobUpdateWorkMode;
  employmentType?: JobUpdateEmploymentType;
}