# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Priva Job Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface Profile {
  id: number;
  userId: number;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  skills?: string[];
  /** @nullable */
  education?: string | null;
  /** @nullable */
  experienceYears?: number | null;
  /** @nullable */
  expectedSalary?: number | null;
  /** @nullable */
  resumeUrl?: string | null;
  user?: User;
}