# lib/api-zod/src/generated/types/savedJob.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Priva Job Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from './job';

export interface SavedJob {
  id: number;
  userId: number;
  jobId: number;
  savedAt: Date;
  job?: Job;
}